Move Server._sys_auth and methods to Tenant
